Readd filename to matching lines
At some point the format of the matching line has been regressed. First
it had the format build_uuid:filename:linenumber:line but then the
filename was dropped. This patch restores the original format.

Also this patch adds a bunch of functional test to avoid such regression
in the future