Fine tune the limit searching
When the --days flag is used the tool makes multiple queries to Zuul to
find the proper limit. However if the guess is pretty close to the
actual limit then the heuristics moves forward slowly. So the tool makes
excessive number of queries with slowly increasing limit. Getting
more builds from Zuul and filtering out old buils is better than doing
many queries. So this patch adds a minimum amount of limit increase for
each query.